Explore a CSV file or a data.frame stored in a Rdata file

Opens the shiny application that allows you to explore a CSV file or a data.frame stored in a Rdata file.

explore(...)

Arguments

...

Additional arguments passed to runApp.

Value

The Shiny application is opened in your browser. To stop the function from running you have to cancel the command using ctrl + c or the equivalent shortcut in your system.

Details

This function runs the Shiny application including in this package. It is basically a wrapper for runApp.

References

For more information on what a Shiny app is check http://shiny.rstudio.com/

Author

Stephen Semick, Leonardo Collado-Torres

Examples

if (FALSE) {
## Open the Shiny application
explore()
## Remember to cancel the command to get back to the R console once you are done analyzing your own data.
}